Fixed Bug from Issue 46
Describe the bug
When the page first loads, the temp is no loading.

To Reproduce
Steps to reproduce the behavior:

Launch the application
all fields are filled in, except the temp in the black circle
Expected behavior
The temp is expected to load upon page load

Screenshots
If applicable, add screenshots to help explain your problem.

Additional context
Add any other context about the problem here.